# src/main/resources/api/api-spec.yaml
openapi: 3.1.0
info:
  title: CRM Lead Management API
  version: 1.0.0
  description: REST API для управления лидами в CRM системе
  contact:
    name: Backend Team
    email: backend@mentee.power
servers:
  - url: http://localhost:8081
    description: Local development server

paths:
  /api/leads:
    get:
      operationId: getLeads
      summary: Получить список всех лидов
      tags:
        - Lead Management
      responses:
        '200':
          description: Успешное получение списка лидов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LeadResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      operationId: createLead
      summary: Создать нового лида
      tags:
        - Lead Management
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LeadFormDto'
      responses:
        '201':
          description: Лид успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeadResponse'
          headers:
            Location:
              schema:
                type: string
              description: URI созданного лида
        '400':
          description: Ошибка валидации входных данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/leads/{id}:
    get:
      operationId: getLeadById
      summary: Получить лида по ID
      tags:
        - Lead Management
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Лид найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeadResponse'
        '404':
          description: Лид не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      operationId: updateLead
      summary: Обновить существующего лида
      tags:
        - Lead Management
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LeadFormDto'
      responses:
        '200':
          description: Лид успешно обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeadResponse'
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Лид не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      operationId: deleteLead
      summary: Удалить лида
      tags:
        - Lead Management
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Лид успешно удалён
        '404':
          description: Лид не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    LeadFormDto:
      type: object
      required:
        - email
        - companyName
        - status
      properties:
        email:
          type: string
          format: email
          example: ivan.petrov@example.com
        companyName:
          type: string
          maxLength: 100
          example: ООО Ромашка
        industry:
          type: string
          example: IT
        status:
          type: string
          enum: [NEW, CONTACTED, QUALIFIED, CONVERTED]
          example: NEW

    LeadResponse:
      type: object
      required:
        - id
        - email
        - status
        - companyName
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        email:
          type: string
          format: email
          example: ivan.petrov@example.com
        status:
          type: string
          enum: [NEW, CONTACTED, QUALIFIED, CONVERTED]
          example: NEW
        companyName:
          type: string
          example: ООО Ромашка
        createdAt:
          type: string
          format: date-time
          example: 2025-11-25T10:30:00Z

    ErrorResponse:
      type: object
      required:
        - timestamp
        - status
        - error
        - message
        - path
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          example: 404
        error:
          type: string
          example: Not Found
        message:
          type: string
          example: Lead with id ... not found
        path:
          type: string
          example: /api/leads/123
        errors:
          type: object
          additionalProperties:
            type: string